% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damage.R
\name{damage}
\alias{damage}
\title{Calculate Damage}
\usage{
damage(power, atk, def, level = 50, modifiers = c())
}
\arguments{
\item{power}{base power of the move being used}

\item{atk}{offensive stat of the attacking Pokémon; for physical moves this is the Attack stat, and for special moves this is the Special Attack stat (with some exceptions)}

\item{def}{defensive stat of the defensive Pokémon; either Defense or Special Defense, similar to atk}

\item{level}{level of the attacking Pokémon (1-100)}

\item{modifiers}{a vector of modifiers affecting the damage calculation}
}
\value{
damage value given as an integer
}
\description{
Calculates the damage inflicted by an attack of given base power, with given
stats, levels, and modifiers.
}
