% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stat_value}
\alias{stat_value}
\title{Calculate Stat Value}
\usage{
stat_value(
  base,
  level = 50,
  iv = 31,
  ev = 252,
  nature = c("positive", "neutral", "negative"),
  is_hp = FALSE
)
}
\arguments{
\item{base}{base stat for the species (1-255)}

\item{level}{level of the Pokémon (1-100)}

\item{iv}{Individual Value in the given stat (0-31)}

\item{ev}{Effort Values in the given stat (0-252)}

\item{nature}{which effect does the Pokémon's nature have on the given stat?}

\item{is_hp}{is the stat being calculated HP or not?}
}
\value{
stat value given as an integer
}
\description{
Calculates the raw stat of a Pokémon, i.e. the numerical value which would be
displayed in the game's summary menu.
}
